hierarchies: API cleanup (#577)
* Move `HierarchyNodeKey` and `HierarchyNodeIdentifier` out of `HierarchyNode.ts` into separate source files

* Cleanup hierarchy definition APIs

* Re-export only `GenericInstanceFilter`, but not its related types

* Export specific errors

* Export less hierarchy node - related stuff

* Export less node key - related stuff

* Export an interface + factory function rather than a class for `HierarchyProvider`

* `IHierarchyLevelDefinitionsFactory` -> `HierarchyLevelDefinitionsFactory`

* Export an interface + factory function instead of `NodeSelectQueryFactory` class.
