+++ apps/full-stack-tests/src/hierarchies/grouping/ClassGrouping.test.ts
@@ -6,7 +6,7 @@
 import { insertPhysicalPartition, insertSubject } from "presentation-test-utilities";
 import { PhysicalPartition, Subject } from "@itwin/core-backend";
 import { IModel } from "@itwin/core-common";
-import { IHierarchyLevelDefinitionsFactory, NodeSelectQueryFactory } from "@itwin/presentation-hierarchies";
+import { createNodesQueryClauseFactory, HierarchyLevelDefinitionsFactory } from "@itwin/presentation-hierarchies";
 import { buildIModel } from "../../IModelUtils";
 import { initialize, terminate } from "../../IntegrationTests";
 import { NodeValidators, validateHierarchy } from "../HierarchyValidation";
@@ -36,8 +36,8 @@ describe("Hierarchies", () => {
         return { childSubject1, childPartition2, childSubject3, childPartition4 };
       });
 
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-      const hierarchy: IHierarchyLevelDefinitionsFactory = {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+      const hierarchy: HierarchyLevelDefinitionsFactory = {
         async defineHierarchyLevel({ parentNode }) {
           if (!parentNode) {
             return [
