+++ apps/full-stack-tests/src/hierarchies/grouping/LabelGrouping.test.ts
@@ -6,7 +6,7 @@
 import { insertSubject } from "presentation-test-utilities";
 import { Subject } from "@itwin/core-backend";
 import { IModel } from "@itwin/core-common";
-import { HierarchyNode, IHierarchyLevelDefinitionsFactory, NodeSelectQueryFactory } from "@itwin/presentation-hierarchies";
+import { createNodesQueryClauseFactory, HierarchyLevelDefinitionsFactory, HierarchyNode } from "@itwin/presentation-hierarchies";
 import { buildIModel } from "../../IModelUtils";
 import { initialize, terminate } from "../../IntegrationTests";
 import { NodeValidators, validateHierarchy } from "../HierarchyValidation";
@@ -36,8 +36,8 @@ describe("Hierarchies", () => {
         return { childSubject1, childSubject2, childSubject3, childSubject4 };
       });
 
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-      const hierarchy: IHierarchyLevelDefinitionsFactory = {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+      const hierarchy: HierarchyLevelDefinitionsFactory = {
         async defineHierarchyLevel(props) {
           if (!props.parentNode) {
             return [
@@ -108,8 +108,8 @@ describe("Hierarchies", () => {
         return { childSubject1, childSubject2, childSubject3, childSubject4 };
       });
 
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-      const hierarchy: IHierarchyLevelDefinitionsFactory = {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+      const hierarchy: HierarchyLevelDefinitionsFactory = {
         async defineHierarchyLevel(props) {
           if (!props.parentNode) {
             return [
@@ -182,8 +182,8 @@ describe("Hierarchies", () => {
         return { rootSubject, childSubject1, childSubject2, childSubject3 };
       });
 
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-      const hierarchy: IHierarchyLevelDefinitionsFactory = {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+      const hierarchy: HierarchyLevelDefinitionsFactory = {
         async defineHierarchyLevel(props) {
           if (!props.parentNode) {
             return [
@@ -235,8 +235,8 @@ describe("Hierarchies", () => {
         return { rootSubject, childSubject1, childSubject2 };
       });
 
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-      const hierarchy: IHierarchyLevelDefinitionsFactory = {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+      const hierarchy: HierarchyLevelDefinitionsFactory = {
         async defineHierarchyLevel(props) {
           if (!props.parentNode) {
             return [
@@ -282,8 +282,8 @@ describe("Hierarchies", () => {
         return { rootSubject, visibleSubject1, visibleSubject2 };
       });
 
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-      const hierarchy: IHierarchyLevelDefinitionsFactory = {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+      const hierarchy: HierarchyLevelDefinitionsFactory = {
         async defineHierarchyLevel(props) {
           if (!props.parentNode) {
             return [
