+++ apps/full-stack-tests/src/hierarchies/grouping/PropertiesGrouping.test.ts
@@ -7,7 +7,7 @@ import { insertSubject } from "presentation-test-utilities";
 import { Subject } from "@itwin/core-backend";
 import { IModel } from "@itwin/core-common";
 import { IModelConnection } from "@itwin/core-frontend";
-import { ECSqlSelectClausePropertiesGroupingParams, IHierarchyLevelDefinitionsFactory, NodeSelectQueryFactory } from "@itwin/presentation-hierarchies";
+import { createNodesQueryClauseFactory, HierarchyLevelDefinitionsFactory, NodesQueryClauseFactory } from "@itwin/presentation-hierarchies";
 import { buildIModel, importSchema, withECDb } from "../../IModelUtils";
 import { initialize, terminate } from "../../IntegrationTests";
 import { NodeValidators, validateHierarchy } from "../HierarchyValidation";
@@ -15,6 +15,9 @@ import { createIModelAccess, createProvider } from "../Utils";
 
 describe("Hierarchies", () => {
   describe("Properties grouping", () => {
+    type ECSqlSelectClausePropertiesGroupingParams = NonNullable<
+      NonNullable<Parameters<NodesQueryClauseFactory["createSelectClause"]>[0]["grouping"]>["byProperties"]
+    >;
     let subjectClassName: string;
 
     before(async function () {
@@ -29,8 +32,8 @@ describe("Hierarchies", () => {
     function createHierarchyWithSpecifiedGrouping(
       imodel: IModelConnection,
       specifiedGrouping: ECSqlSelectClausePropertiesGroupingParams,
-    ): IHierarchyLevelDefinitionsFactory {
-      const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
+    ): HierarchyLevelDefinitionsFactory {
+      const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
       return {
         async defineHierarchyLevel({ parentNode }) {
           if (!parentNode) {
@@ -179,8 +182,8 @@ describe("Hierarchies", () => {
           return { childSubject1, childSubject2 };
         });
 
-        const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-        const customHierarchy: IHierarchyLevelDefinitionsFactory = {
+        const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+        const customHierarchy: HierarchyLevelDefinitionsFactory = {
           async defineHierarchyLevel({ parentNode }) {
             if (!parentNode) {
               return [
@@ -308,8 +311,8 @@ describe("Hierarchies", () => {
             return { schema, x1, x2 };
           },
           async (imodel, { schema, x1, x2 }) => {
-            const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-            const hierarchy: IHierarchyLevelDefinitionsFactory = {
+            const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+            const hierarchy: HierarchyLevelDefinitionsFactory = {
               async defineHierarchyLevel({ parentNode }) {
                 if (!parentNode) {
                   return [
@@ -373,8 +376,8 @@ describe("Hierarchies", () => {
             return { schema, x1 };
           },
           async (imodel, { schema, x1 }) => {
-            const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-            const hierarchy: IHierarchyLevelDefinitionsFactory = {
+            const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+            const hierarchy: HierarchyLevelDefinitionsFactory = {
               async defineHierarchyLevel({ parentNode }) {
                 if (!parentNode) {
                   return [
@@ -441,8 +444,8 @@ describe("Hierarchies", () => {
             return { schema, x1, x2 };
           },
           async (imodel, { schema, x1, x2 }) => {
-            const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-            const hierarchy: IHierarchyLevelDefinitionsFactory = {
+            const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+            const hierarchy: HierarchyLevelDefinitionsFactory = {
               async defineHierarchyLevel({ parentNode }) {
                 if (!parentNode) {
                   return [
@@ -524,8 +527,8 @@ describe("Hierarchies", () => {
             return { schema, x1 };
           },
           async (imodel, { schema, x1 }) => {
-            const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-            const hierarchy: IHierarchyLevelDefinitionsFactory = {
+            const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+            const hierarchy: HierarchyLevelDefinitionsFactory = {
               async defineHierarchyLevel({ parentNode }) {
                 if (!parentNode) {
                   return [
@@ -583,8 +586,8 @@ describe("Hierarchies", () => {
             return { schema, x1, x2 };
           },
           async (imodel, { schema, x1, x2 }) => {
-            const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-            const hierarchy: IHierarchyLevelDefinitionsFactory = {
+            const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+            const hierarchy: HierarchyLevelDefinitionsFactory = {
               async defineHierarchyLevel({ parentNode }) {
                 if (!parentNode) {
                   return [
@@ -652,8 +655,8 @@ describe("Hierarchies", () => {
             return { schema, x, y };
           },
           async (imodel, { schema, x, y }) => {
-            const selectQueryFactory = new NodeSelectQueryFactory({ imodelAccess: createIModelAccess(imodel) });
-            const hierarchy: IHierarchyLevelDefinitionsFactory = {
+            const selectQueryFactory = createNodesQueryClauseFactory({ imodelAccess: createIModelAccess(imodel) });
+            const hierarchy: HierarchyLevelDefinitionsFactory = {
               async defineHierarchyLevel({ parentNode }) {
                 if (!parentNode) {
                   return [
