+++ apps/performance-tests/src/hierarchies/HideIfNoChildren.test.ts
@@ -6,10 +6,11 @@
 import { expect } from "chai";
 import { IModelDb, SnapshotDb } from "@itwin/core-backend";
 import {
-  ClassBasedHierarchyLevelDefinitionsFactory,
+  createClassBasedHierarchyLevelDefinitionsFactory,
+  createNodesQueryClauseFactory,
   DefineInstanceNodeChildHierarchyLevelProps,
   HierarchyLevelDefinition,
-  NodeSelectQueryFactory,
+  NodesQueryClauseFactory,
 } from "@itwin/presentation-hierarchies";
 import { Datasets } from "../util/Datasets";
 import { run } from "../util/TestUtilities";
@@ -27,8 +28,8 @@ describe("hide if no children", () => {
       const provider = new StatelessHierarchyProvider({
         iModel,
         getHierarchyFactory: (imodelAccess) => {
-          const queryFactory = new NodeSelectQueryFactory({ imodelAccess });
-          return new ClassBasedHierarchyLevelDefinitionsFactory({
+          const queryFactory = createNodesQueryClauseFactory({ imodelAccess });
+          return createClassBasedHierarchyLevelDefinitionsFactory({
             classHierarchyInspector: imodelAccess,
             hierarchy: {
               rootNodes: async () => createPhysicalElementsHierarchyLevelDefinition({ queryFactory, limit: 5 }),
@@ -64,8 +65,8 @@ describe("hide if no children", () => {
       const provider = new StatelessHierarchyProvider({
         iModel,
         getHierarchyFactory: (imodelAccess) => {
-          const queryFactory = new NodeSelectQueryFactory({ imodelAccess });
-          return new ClassBasedHierarchyLevelDefinitionsFactory({
+          const queryFactory = createNodesQueryClauseFactory({ imodelAccess });
+          return createClassBasedHierarchyLevelDefinitionsFactory({
             classHierarchyInspector: imodelAccess,
             hierarchy: {
               rootNodes: async () => createPhysicalElementsHierarchyLevelDefinition({ queryFactory, limit: 5 }),
@@ -95,7 +96,7 @@ describe("hide if no children", () => {
 });
 
 async function createPhysicalElementsHierarchyLevelDefinition(props: {
-  queryFactory: NodeSelectQueryFactory;
+  queryFactory: NodesQueryClauseFactory;
   limit: number;
   hasChildren?: boolean;
   hideIfNoChildren?: boolean;
