+++ packages/hierarchies/api/presentation-hierarchies.exports.csv
@@ -1,73 +1,37 @@
 sep=;
 Release Tag;API Item
-beta;BaseHierarchyNode
-beta;ClassBasedHierarchyDefinition
-beta;ClassBasedHierarchyDefinitionsFactoryProps
-beta;ClassBasedHierarchyLevelDefinition = InstancesNodeChildHierarchyLevelDefinition | CustomNodeChildHierarchyLevelDefinition
-beta;ClassBasedHierarchyLevelDefinitionsFactory 
-beta;ClassGroupingNodeKey
+beta;createClassBasedHierarchyLevelDefinitionsFactory(props: ClassBasedHierarchyDefinitionsFactoryProps): HierarchyLevelDefinitionsFactory
+beta;createHierarchyProvider(props: HierarchyProviderProps): HierarchyProvider
 beta;createLimitingECSqlQueryExecutor(baseExecutor: ECSqlQueryExecutor, defaultLimit: number | "unbounded"): LimitingECSqlQueryExecutor
-beta;CustomHierarchyNodeDefinition
-beta;CustomNodeChildHierarchyLevelDefinition
+beta;createNodesQueryClauseFactory(props:
+beta;DefineCustomNodeChildHierarchyLevelProps = Omit
 beta;DefineHierarchyLevelProps
-beta;DefineInstanceNodeChildHierarchyLevelProps = DefineHierarchyLevelProps &
+beta;DefineInstanceNodeChildHierarchyLevelProps = Omit
 beta;DefineRootHierarchyLevelProps = Omit
-beta;ECSqlSelectClauseBaseClassGroupingParams 
-beta;ECSqlSelectClauseGroupingParams
-beta;ECSqlSelectClauseGroupingParamsBase
-beta;ECSqlSelectClauseLabelGroupingBaseParams
-beta;ECSqlSelectClauseLabelGroupingGroupParams 
-beta;ECSqlSelectClauseLabelGroupingMergeParams 
-beta;ECSqlSelectClauseLabelGroupingParams = boolean | ECSqlValueSelector | ECSqlSelectClauseLabelGroupingMergeParams | ECSqlSelectClauseLabelGroupingGroupParams
-beta;ECSqlSelectClausePropertiesGroupingParams 
-beta;ECSqlSelectClausePropertyGroup
-beta;ECSqlSelectClausePropertyValueRange
-beta;ECSqlValueSelector
 beta;GetHierarchyNodesProps
 beta;getLogger(): ILogger
 beta;GroupingHierarchyNode 
 beta;GroupingNodeKey = ClassGroupingNodeKey | LabelGroupingNodeKey | PropertyGroupingNodeKey
-beta;HierarchyDefinitionParentNode = Omit
 beta;HierarchyLevelDefinition = HierarchyNodesDefinition[]
+beta;HierarchyLevelDefinitionsFactory
 beta;HierarchyNode = NonGroupingHierarchyNode | GroupingHierarchyNode
 beta;HierarchyNode
-beta;HierarchyNodeAutoExpandProp = "single-child" | "always"
-beta;HierarchyNodeBaseClassGroupingParams 
-beta;HierarchyNodeGroupingParams
-beta;HierarchyNodeGroupingParamsBase
 beta;HierarchyNodeIdentifier = InstanceKey |
 beta;HierarchyNodeIdentifier
 beta;HierarchyNodeIdentifiersPath = HierarchyNodeIdentifier[]
 beta;HierarchyNodeKey = StandardHierarchyNodeKey | string
 beta;HierarchyNodeKey
-beta;HierarchyNodeLabelGroupingBaseParams
-beta;HierarchyNodeLabelGroupingGroupParams 
-beta;HierarchyNodeLabelGroupingMergeParams 
-beta;HierarchyNodeLabelGroupingParams = boolean | HierarchyNodeLabelGroupingMergeParams | HierarchyNodeLabelGroupingGroupParams
-beta;HierarchyNodeProcessingParamsBase
-beta;HierarchyNodePropertiesGroupingParams 
-beta;HierarchyNodePropertyGroup
-beta;HierarchyNodePropertyValueRange
 beta;HierarchyNodesDefinition = CustomHierarchyNodeDefinition | InstanceNodesQueryDefinition
 beta;HierarchyNodesDefinition
 beta;HierarchyProvider
-beta;HierarchyProviderLocalizedStrings
-beta;HierarchyProviderProps
-beta;IHierarchyLevelDefinitionsFactory
-beta;INodeParser = (row:
-beta;INodePostProcessor = (node: ProcessedHierarchyNode) => Promise
-beta;INodePreProcessor = 
-beta;InstanceHierarchyNodeProcessingParams 
-beta;InstanceNodesQueryDefinition
-beta;InstancesNodeChildHierarchyLevelDefinition
 beta;InstancesNodeKey
-beta;LabelGroupingNodeKey
 beta;LimitingECSqlQueryExecutor
+beta;NodeParser = (row:
+beta;NodePostProcessor = (node: ProcessedHierarchyNode) => Promise
+beta;NodePreProcessor = 
 beta;NodeSelectClauseColumnNames
-beta;NodeSelectClauseProps
-beta;NodeSelectQueryFactory
+beta;NodesQueryClauseFactory
 beta;NonGroupingHierarchyNode 
-beta;ParentHierarchyNode
 beta;ParsedCustomHierarchyNode = Omit
 beta;ParsedHierarchyNode = ParsedCustomHierarchyNode | ParsedInstanceHierarchyNode
 beta;ParsedInstanceHierarchyNode = Omit
@@ -75,10 +39,5 @@ beta;ProcessedCustomHierarchyNode = Omit
 beta;ProcessedGroupingHierarchyNode = Omit
 beta;ProcessedHierarchyNode = ProcessedCustomHierarchyNode | ProcessedInstanceHierarchyNode | ProcessedGroupingHierarchyNode
 beta;ProcessedInstanceHierarchyNode = Omit
-beta;PropertyGroupingNodeKey = PropertyValueRangeGroupingNodeKey | PropertyValueGroupingNodeKey | PropertyOtherValuesGroupingNodeKey
-beta;PropertyOtherValuesGroupingNodeKey
-beta;PropertyValueGroupingNodeKey
-beta;PropertyValueRangeGroupingNodeKey
 beta;RowsLimitExceededError 
 beta;setLogger(logger: ILogger | undefined): void
-beta;StandardHierarchyNodeKey = InstancesNodeKey | GroupingNodeKey
