+++ packages/hierarchies/src/hierarchies/internal/FilteringHierarchyLevelDefinitionsFactory.ts
@@ -8,27 +8,21 @@ import {
   CustomHierarchyNodeDefinition,
   DefineHierarchyLevelProps,
   HierarchyLevelDefinition,
+  HierarchyLevelDefinitionsFactory,
   HierarchyNodesDefinition,
-  IHierarchyLevelDefinitionsFactory,
-  INodeParser,
-  INodePostProcessor,
-  INodePreProcessor,
   InstanceNodesQueryDefinition,
+  NodeParser,
+  NodePostProcessor,
+  NodePreProcessor,
 } from "../HierarchyDefinition";
-import {
-  HierarchyNode,
-  HierarchyNodeIdentifier,
-  HierarchyNodeIdentifiersPath,
-  ParsedHierarchyNode,
-  ParsedInstanceHierarchyNode,
-  ProcessedHierarchyNode,
-} from "../HierarchyNode";
+import { HierarchyNode, ParsedHierarchyNode, ParsedInstanceHierarchyNode, ProcessedHierarchyNode } from "../HierarchyNode";
+import { HierarchyNodeIdentifier, HierarchyNodeIdentifiersPath } from "../HierarchyNodeIdentifier";
 import { defaultNodesParser } from "./TreeNodesReader";
 
 /** @internal */
 export interface FilteringQueryBuilderProps {
   classHierarchy: ECClassHierarchyInspector;
-  source: IHierarchyLevelDefinitionsFactory;
+  source: HierarchyLevelDefinitionsFactory;
   nodeIdentifierPaths: HierarchyNodeIdentifiersPath[];
 }
 
@@ -39,9 +33,9 @@ export type FilteredHierarchyNode<TNode = ProcessedHierarchyNode> = TNode & {
 };
 
 /** @internal */
-export class FilteringHierarchyLevelDefinitionsFactory implements IHierarchyLevelDefinitionsFactory {
+export class FilteringHierarchyLevelDefinitionsFactory implements HierarchyLevelDefinitionsFactory {
   private _classHierarchy: ECClassHierarchyInspector;
-  private _source: IHierarchyLevelDefinitionsFactory;
+  private _source: HierarchyLevelDefinitionsFactory;
   private _nodeIdentifierPaths: HierarchyNodeIdentifiersPath[];
 
   public constructor(props: FilteringQueryBuilderProps) {
@@ -50,7 +44,7 @@ export class FilteringHierarchyLevelDefinitionsFactory implements IHierarchyLeve
     this._nodeIdentifierPaths = props.nodeIdentifierPaths;
   }
 
-  public get preProcessNode(): INodePreProcessor {
+  public get preProcessNode(): NodePreProcessor {
     return async (node) => {
       const processedNode = this._source.preProcessNode ? await this._source.preProcessNode(node) : node;
       if (processedNode?.processingParams?.hideInHierarchy && (node as FilteredHierarchyNode).isFilterTarget) {
@@ -61,7 +55,7 @@ export class FilteringHierarchyLevelDefinitionsFactory implements IHierarchyLeve
     };
   }
 
-  public get postProcessNode(): INodePostProcessor {
+  public get postProcessNode(): NodePostProcessor {
     return async (node: ProcessedHierarchyNode) => {
       const processedNode = this._source.postProcessNode ? await this._source.postProcessNode(node) : node;
       if (
@@ -75,7 +69,7 @@ export class FilteringHierarchyLevelDefinitionsFactory implements IHierarchyLeve
     };
   }
 
-  public get parseNode(): INodeParser {
+  public get parseNode(): NodeParser {
     return (row: { [columnName: string]: any }): FilteredHierarchyNode<ParsedInstanceHierarchyNode> => {
       const isFilterTarget: boolean = !!row[ECSQL_COLUMN_NAME_IsFilterTarget];
       const parsedFilteredChildrenIdentifierPaths: HierarchyNodeIdentifiersPath[] | undefined = row[ECSQL_COLUMN_NAME_FilteredChildrenPaths]
@@ -86,9 +80,11 @@ export class FilteringHierarchyLevelDefinitionsFactory implements IHierarchyLeve
     };
   }
 
-  public async defineHierarchyLevel(props: DefineHierarchyLevelProps): Promise<HierarchyLevelDefinition> {
+  public async defineHierarchyLevel(
+    props: DefineHierarchyLevelProps & { parentNode: FilteredHierarchyNode<DefineHierarchyLevelProps["parentNode"]> | undefined },
+  ): Promise<HierarchyLevelDefinition> {
     const sourceDefinitions = await this._source.defineHierarchyLevel(props);
-    const { filteredNodePaths, isParentFilterTarget } = this.getFilteringProps(props.parentNode as FilteredHierarchyNode | undefined);
+    const { filteredNodePaths, isParentFilterTarget } = this.getFilteringProps(props.parentNode);
     if (!filteredNodePaths) {
       return sourceDefinitions;
     }
@@ -139,7 +135,7 @@ export class FilteringHierarchyLevelDefinitionsFactory implements IHierarchyLeve
     return filteredDefinitions;
   }
 
-  private getFilteringProps(parentNode: FilteredHierarchyNode | undefined) {
+  private getFilteringProps(parentNode: FilteredHierarchyNode<unknown> | undefined) {
     if (!parentNode) {
       return { filteredNodePaths: this._nodeIdentifierPaths, isParentFilterTarget: false };
     }
