+++ packages/hierarchies/src/hierarchies/internal/operators/grouping/BaseClassGrouping.ts
@@ -4,7 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { createMainThreadReleaseOnTimePassedHandler, EC, ECClassHierarchyInspector, ECSchemaProvider, getClass } from "@itwin/presentation-shared";
-import { ClassGroupingNodeKey, HierarchyNode, ParentHierarchyNode, ProcessedInstanceHierarchyNode } from "../../../HierarchyNode";
+import { HierarchyNode, ParentHierarchyNode, ProcessedInstanceHierarchyNode } from "../../../HierarchyNode";
+import { ClassGroupingNodeKey } from "../../../HierarchyNodeKey";
 import { GroupingHandler, GroupingHandlerResult } from "../Grouping";
 
 /** @internal */
