+++ packages/hierarchies/src/test/HierarchyNodeIdentifier.test.ts
@@ -1,0 +1,49 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+import { expect } from "chai";
+import { InstanceKey } from "@itwin/presentation-shared";
+import { HierarchyNodeIdentifier } from "../hierarchies/HierarchyNodeIdentifier";
+
+describe("HierarchyNodeIdentifier", () => {
+  const instanceNodeIdentifier: InstanceKey = {
+    className: "a",
+    id: "0x1",
+  };
+  const customNodeIdentifier = {
+    key: "x",
+  };
+
+  describe("isInstanceNodeIdentifier", () => {
+    it("returns correct result for different types of identifiers", () => {
+      expect(HierarchyNodeIdentifier.isInstanceNodeIdentifier(instanceNodeIdentifier)).to.be.true;
+      expect(HierarchyNodeIdentifier.isInstanceNodeIdentifier(customNodeIdentifier)).to.be.false;
+    });
+  });
+
+  describe("isCustomNodeIdentifier", () => {
+    it("returns correct result for different types of identifiers", () => {
+      expect(HierarchyNodeIdentifier.isCustomNodeIdentifier(instanceNodeIdentifier)).to.be.false;
+      expect(HierarchyNodeIdentifier.isCustomNodeIdentifier(customNodeIdentifier)).to.be.true;
+    });
+  });
+
+  describe("equal", () => {
+    it("compares custom node identifiers", () => {
+      expect(HierarchyNodeIdentifier.equal(customNodeIdentifier, { key: "y" })).to.be.false;
+      expect(HierarchyNodeIdentifier.equal(customNodeIdentifier, { key: "x" })).to.be.true;
+    });
+
+    it("compares instance node identifiers", () => {
+      expect(HierarchyNodeIdentifier.equal(instanceNodeIdentifier, { className: "b", id: "0x1" })).to.be.false;
+      expect(HierarchyNodeIdentifier.equal(instanceNodeIdentifier, { className: "a", id: "0x2" })).to.be.false;
+      expect(HierarchyNodeIdentifier.equal(instanceNodeIdentifier, { className: "a", id: "0x1" })).to.be.true;
+    });
+
+    it("compares instance and custom node identifiers", () => {
+      expect(HierarchyNodeIdentifier.equal(instanceNodeIdentifier, customNodeIdentifier)).to.be.false;
+    });
+  });
+});
